🐛 Add `Loader` to loading state of User Profile page
Closes #164
