+++ packages/ui/src/modules/user/components/Account/Profile/component.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as MutationHooks from 'shared/graphql/mutation/useUserRegister';
 import { render } from 'testHelpers';
-import { Profile } from '.';
+import { Profile } from './component';
 
 jest.mock('shared/Content');
 jest.mock('shared/Page');
