+++ packages/ui/src/modules/user/components/Account/Security/index.js
@@ -1,40 +1,11 @@
-import { always, equals } from 'ramda';
-import React from 'react';
-import { useLocation } from 'react-router-dom';
-import { Body, Breadcrumb, Breadcrumbs, Header } from 'shared/Content';
-import { useTranslate } from 'shared/useTranslate';
-import { Route } from 'types/route';
-import { Form } from './Form';
+import { Utils } from '@boilerplate-monorepo/common';
+import { lazy } from 'react';
 
-const Security = () => {
-  const t = useTranslate({
-    component: 'user',
-    namespace: 'user',
-  });
+const Lazy = lazy(() =>
+  import(
+    /* webpackChunkName: "userSecurity" */
+    './component'
+  ).then(Utils.renameKeys({ Security: 'default' }))
+);
 
-  const location = useLocation();
-  const isActive = equals(location.pathname, Route.USER_ACCOUNT.path);
-
-  return (
-    <>
-      <Breadcrumbs>
-        <Breadcrumb
-          isActive={always(isActive)}
-          text={t('account')}
-          to={Route.USER_ACCOUNT.path}
-        />
-        <Breadcrumb
-          isEnd
-          text={t('security')}
-          to={Route.USER_ACCOUNT_SECURITY.path}
-        />
-      </Breadcrumbs>
-      <Header title={t('security')} />
-      <Body>
-        <Form />
-      </Body>
-    </>
-  );
-};
-
-export { Security };
+export { Lazy as Security };
